Re-structure project folders
Create a folder for each PEC and move the specific folders and files into the correspondent folder. Bibliography file is common to all PECs and stays at floor level.

The goal of this changes is to keep a clean folder structure for an easier control version. Since the TFM is a unique project I feel that it makes sense to commit and push all its different parts (PEC) at the same time.