Add Romina correction/suggestion on tokenization
I have used 'itemizacion' as a translation of tokenization, but that
is not correct. I have used its definition, instead, but conserved the
word 'token' because it is commonly used in Spanish in computer jargon.

Another change has been changing the word 'item' along the text with the
word 'token'.