Extract abstracts using query
Submit a query to PubMed and retry the resulting registers,
save as text files and make the info into an S4 object
of class 'Abstracts' for using with package pubmed.mineR.

New file containing the code for making and processing the
objects. Saves big objects, and those with require long
processing time as RData files. These will be loaded by
main text document when necessary for knitting the report.

Added directory "data/" to gitignore, since those files are
created by running the code.