Check specificity of gene_atomization and pubtator_function for AMH gene
Uses pubtator_function on a pmid whose abstract contains reference
to AMH as hormone, no gene. The result states that it makes reference
to the gene, pointing the for this case at least, pubtator_function
is not more specific than gene_atomization.

With official_fn function it was clear that abbreviation AMH for
anti-mullerian hormone was detected as the HGNC symbol for a
corresponding gene.