Read gene2pubtatorcentral into a dataframe
Read the text file gene2pubtatorcentral and saves pmid and
geneID fields into a dataframe. Saves the dataframe as an
RData file in intermediateData directory.

gene2pubtatorcentral contains info about different entities in
the abstracts of PubMed databases cites. I want to recover the
data corresponding to GeneID and its frequencies. The first step
is reading and loading only the columns that correspond to pmid and
geneID, and saving as a dataframe. Next step is cleaning instances
in which GeneID field contains 2 or more codes. Tricky memory-wise,
since the resulting dataframe contains more than 50 million rows.