Add frequency data of geneID from PubTator
The file PEC2/PEC2_fase1_code.Rmd contains the code for
loading and cleaning the necessary dataset, and for summarizing
it into a contingency table.

The resulting contingency table is stored as an RData file
in data/ folder. It should be considered as immutable data.

The Shiny app will use the contingency table in the
hypergeometric test filter.