Create file PEC3_fase2_script.R
It is just a copy of the file PEC3_fase2_script.R.
This one will develop the former with the features en PEC3. Firstly I
was everything in the PEC2 version, but I want to have clear which
deliverables go with each PEC. Also, less complexity in the addresses for
reading and writing files.