Simplify enrichResult objects
Uses function simplify() from package clusterProfiler for getting rid of redundant GO terms.

The necessary time for performing the operation grows exponentially with the number of GO terms in the object. I made a rough calculation for introducing warnings to the user (1500, 2500, and 4300 GO terms as milestones for slightly long, uncomfortably long, maybe too long for waiting).