Compute enrichment of GO terms with Enrichr through package enrichR
Using package enrichR instead of package clusterProfiler.
At the point of this commit, only for CC GO terms and no barplot.

The webapp at the shinyapp.io server froze whenever tried to compute
enriched GO terms with the function enrichGO. I haven't found any way
to prevent it.

The only (free) solution I have found is to outsource the computing
to the web app Enrichr, using the package enrichR as an interface for
their API.
I would prefer using enrichGO because I understand better what it
does and what the results mean, and you have more control over it
but it can not be helped.