Set GO terms table for user to download
The table prepared for downloading contains more info that
the one shown in the app: It is not cut off by adjusted p-value;
retains columns Term, p-value, adjusted-p-value, Score.combined, Overlap,
and Genes. The column names are in Spanish, as the ones shown in the app.

The reason for doing so is giving more choices to the user for,
f.ex. using original p-values for adjusting them with his
preferred method, extracting the genes in each term, etc.